Add more scalb test expectations for "inexact" exception.
This patch adds more libm-test.inc expectations for the "inexact"
exception for scalb, in all cases except those with a non-integer
second argument (where results are unspecified by POSIX, so the
function does not count as fully determined and the spurious "inexact"
exceptions raised by the existing implementations alongside "invalid"
are OK).

Tested for x86_64 and x86.

	* math/libm-test.inc (scalb_test_data): Add more expectations for
	the "inexact" exception.