Correct "inexact" expectations in lround, llround tests.
I noticed that some of my recently added tests of lround and llround
wrongly expected the "inexact" exception to be absent for certain
within-range non-integer arguments.  (It's unspecified whether this
exception is present or not for within-range non-integer arguments; it
mustn't be present for integer arguments and out-of-range arguments.)
This patch corrects those expectations.

Tested for x86_64 and x86.

	* math/libm-test.inc (lround_test_data): Do not expect the absence
	of "inexact" for some tests with non-integer arguments.
	(llround_test_data): Likewise.