Add more tests for ceil, floor, round, trunc.
This patch adds more tests for ceil, floor, round and trunc, with a
particular focus on verifying they don't raise spurious "inexact"
exceptions for integer arguments (a C99 / C11 requirement, as opposed
to the general principle that they shouldn't raise "inexact" for any
arguments at all which is a TS 18661-1 requirement).

Tested for x86_64 and x86.

	* math/libm-test.inc (ceil_test_data): Add more tests and more
	expectations for "inexact".
	(floor_test_data): Add more tests.
	(round_test_data): Likewise.
	(trunc_test_data): Likewise.