Remove libm-test.inc special-casing of errors up to 0.5 ulp.
libm-test.inc has special-case code treating errors of up to 0.5 ulp
as allowed (for functions that aren't exactly determined) even if no
such errors appeared in libm-test-ulps.  This only applies to avoid
errors for individual function calls, not for the overall check of
ulps at the end of testing a function, resulting in confusing output
of the form:

testing double (without inline functions)
Maximal error of `log_upward'
 is      : 1 ulp
 accepted: 0 ulp

with no report of what testcase produced that error.  This patch
removes the special case, so that instead you get:

testing double (without inline functions)
Failure: Test: log_upward (0x1.0000000000001p+0)
Result:
 is:          2.2204460492503129e-16   0x1.fffffffffffffp-53
 should be:   2.2204460492503131e-16   0x1.0000000000000p-52
 difference:  2.4651903288156619e-32   0x1.0000000000000p-105
 ulp       :  0.5000
 max.ulp   :  0.0000
Maximal error of `log_upward'
 is      : 1 ulp
 accepted: 0 ulp

(for formats other than ldbl-128ibm, 0.5 ulp errors only occur in
unusual cases such as this where the correctly rounded result is a
power of 2 and the computed result is just below it).  This should not
affect which cases result in the test failing, just ensure that if it
fails then some failure for an individual function call was reported.

Tested for x86_64 and x86.

	* math/libm-test.inc (check_float_internal): Do not special-case
	errors up to 0.5 ulp.