Remove configure tests for visibility support.
There are various configure tests for visibility support in the
compiler and assember.

GCC support for visibility attributes was added in GCC 3.3.  I don't
know what specific fix was intended by the test "for broken
__attribute__((visibility())", but it was added by
<https://sourceware.org/ml/libc-hacker/2002-08/msg00030.html>, and GCC
3.3 appears not to have that breakage, so I suspect it was only ever
in development versions before 3.3 was released.  The assembler
support was added in binutils 2.10.

This patch removes the tests in question as obsolete.  Two tests that
were formerly conditional on "if test $libc_cv_visibility_attribute =
yes", including the one for linker support for protected data, are now
unconditional.

Tested for x86_64 and x86 (testsuite, and that installed shared
libraries are unchanged by the patch).

	* configure.ac (libc_cv_asm_protected_directive): Remove configure
	test.
	(libc_cv_visibility_attribute): Likewise.
	(libc_cv_protected_data): Test unconditionally.
	(libc_cv_broken_visibility_attribute): Remove configure test.
	(libc_cv_have_sdata_section): Test unconditionally.
	* configure: Regenerated.