Do not test sign of zero result from infinite argument to Bessel functions.
When Bessel functions return a zero result from an infinite argument,
the function oscillates as it approaches 0, so the sign of that zero
result should be indeterminate.  This patch weakens the expectations
accordingly not to check the sign of such results (the tests were
causing spurious failures for j1 (-Inf) for ldbl-128).

Tested for x86_64 and x86.

	* math/libm-test.inc (j0_test_data): Do not test sign of zero
	result from infinite argument.
	(j1_test_data): Likewise.
	(jn_test_data): Likewise.
	(y0_test_data): Likewise.
	(y1_test_data): Likewise.
	(yn_test_data): Likewise.