Refactor libm-test inline tests disabling.
This patch refactors how libm-test.inc handles disabling
errno/exception handling tests, and some other tests, for
__FAST_MATH__ inline function tests.

The macro TEST_INLINE is changed from being defined/undefined to being
defined to 1 or 0, so that it can be tested in "if" conditionals
instead of #if.  For tests of errno and exception setting, separate
macros TEST_ERRNO and TEST_EXCEPTIONS are added, and TEST_ERRNO is
also used in the disabling of errno testing for vector function
tests.  The relevant conditionals are moved up a function, so that
they take effect before the code that counts the number of tests, so
the inline function tests now accurately report that 0 tests for
exceptions and errno were executed (whereas they previously reported a
large number desipte not running any such tests).

Tested for x86_64 and x86.

	* math/test-math-errno.h: New file.
	* math/test-math-inline.h (TEST_INLINE): Define to 1 instead of
	empty.
	(TEST_ERRNO): New macro.
	(TEST_EXCEPTIONS): Likewise.
	* math/test-math-no-inline.h (TEST_INLINE): Likewise.
	(TEST_EXCEPTIONS): Likewise.
	* math/test-math-vector.h (TEST_ERRNO): Likewise.
	* math/test-double.c: Include "test-math-errno.h".
	* math/test-float.c: Likewise.
	* math/test-ldouble.c: Likewise.
	* math/libm-test.inc (test_single_exception) [!TEST_INLINE]: Make
	code unconditional.
	(test_exceptions): Only run code if TEST_EXCEPTIONS.
	(test_single_errno) [!TEST_INLINE && !TEST_MATHVEC]: Make code
	unconditional.
	(test_errno): Only run code if TEST_ERRNO.
	(enable_test): Use "if" conditional on TEST_INLINE, not #ifdef.